+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 # Changelog
 
 All notable changes to this project will be documented in this file. 
-The format is based on [keepachangelog.com]
+The format base on [keepachangelog.com]
 
-### Unreleased
+## [0.0.2] 2021-10-10
 
 ### Changed
 
@@ -26,4 +26,6 @@ The format is based on [keepachangelog.com]
 
 [keepachangelog.com]:https://keepachangelog.com/en/1.0.0/
 [test API]:https://github.com/NaMax66/test-api
+
 [0.0.1]: https://github.com/NaMax66/web-app-architecture-pattern/releases/tag/v0.0.1
+[0.0.2]: https://github.com/NaMax66/web-app-architecture-pattern/compare/v0.0.1...v0.0.2